+++ NEWS.md
@@ -1,3 +1,5 @@
+# snowflakeauth (development version)
+
 # snowflakeauth 0.2.2
 
 * [On-disk connection caching][id_tokens] in the system keyring is now